# swagger.yaml
swagger: "2.0"
info:
  description: "This is a swagger spec for APIs provided by superdentist.io"
  version: "1.0.0"
  title: "SuperDentist APIs"
  contact:
    email: "superdentist.admin@superdentist.io"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "superdentist.io"
basePath: "/api/sd"
...
schemes:
- "https"
- "wss"
paths:
...
  /v1/referral/delete/{id}:
    delete:
      tags:
       - "Referrals"
      summary: "Create a referral from GD to Specialist"
      description: "Create a Referral and send out emails"
      operationId: "DeleteReferral"
      produces:
      - "application/json"
      parameters:
      - in: "path"
        type: string
        name: "id"
        description: "unique address id of registered clinic"
        required: true
      responses:
        '200':
          description: "Referral successfully deleted"
        '401':
          description: "Unauthorized: Bad request or authorization details"
        '500':
          description: "Internal server error occured"
      security:
        - Bearer: []
  /v1/referral/addComments/{id}:
    post:
      tags:
       - "Referrals"
      summary: "Add comments to existing referral"
      description: "Add comments to referral"
      operationId: "AddCommentsToReferral"
      produces:
      - "application/json"
      parameters:
      - in: "path"
        type: string
        name: "id"
        description: "unique address id of registered clinic"
        required: true
      - in: "body"
        name: "body"
        description: "Created clinic object"
        required: true
        schema:
          $ref: "#/definitions/ReferralComments"
      responses:
        '200':
          $ref: '#/responses/CreateReferralResponse'
        '401':
          description: "Unauthorized: Bad request or authorization details"
        '500':
          description: "Internal server error occured"
      security:
        - Bearer: []
  /v1/referral/updateStatus/{id}:
    post:
      tags:
       - "Referrals"
      summary: "Update status of a referral"
      description: "Update status of a referral"
      operationId: "UpdateReferralStatus"
      produces:
      - "application/json"
      parameters:
      - in: "path"
        type: string
        name: "id"
        description: "unique address id of registered clinic"
        required: true
      - in: "body"
        name: "body"
        description: "Created clinic object"
        required: true
        schema:
          $ref: "#/definitions/ReferralStatus"
      responses:
        '200':
          $ref: '#/responses/CreateReferralResponse'
        '401':
          description: "Unauthorized: Bad request or authorization details"
        '500':
          description: "Internal server error occured"
      security:
        - Bearer: []
  /v1/referral/uploadDocuments/{id}:
    post:
      tags:
       - "Referrals"
      summary: "Upload supporting document files"
      description: "Upload supporting documents"
      operationId: "UploadDocuments"
      consumes:
         - multipart/form-data
      produces:
      - "application/json"
      parameters:
      - in: "path"
        type: string
        name: "id"
        description: "unique address id of registered clinic"
        required: true
      responses:
        '200':
          $ref: '#/responses/CreateReferralResponse'
        '401':
          description: "Unauthorized: Bad request or authorization details"
        '500':
          description: "Internal server error occured"
      security:
        - Bearer: []
  /v1/referral/gdReferrals:
    get:
      tags:
       - "Referrals"
      summary: "Get all referrall for GD with addressId"
      description: "Get all referrall for GD with addressId"
      operationId: "GetAllReferralsGD"
      consumes:
         - application/json
      produces:
      - "application/json"
      parameters:
      - in: "query"
        name: "addressId"
        type: string
        required: true
      responses:
        '200':
          $ref: '#/responses/GetReferralListResponse'
        '401':
          description: "Unauthorized: Bad request or authorization details"
        '500':
          description: "Internal server error occured"
      security:
        - Bearer: []
  /v1/referral/spReferrals:
    get:
      tags:
       - "Referrals"
      summary: "Get all referrall for SP with addressId"
      description: "Get all referrall for SP with addressId"
      operationId: "GetAllReferralsSP"
      consumes:
         - application/json
      produces:
      - "application/json"
      parameters:
      - in: "query"
        name: "addressId"
        type: string
        required: true
      responses:
        '200':
          $ref: '#/responses/GetReferralListResponse'
        '401':
          description: "Unauthorized: Bad request or authorization details"
        '500':
          description: "Internal server error occured"
      security:
        - Bearer: []
  /v1/referral/getOne/{id}:
    get:
      tags:
       - "Referrals"
      summary: "Get referral with given referral id"
      description:  "Get referral with given referral id"
      operationId: "GetOneReferral"
      consumes:
         - application/json
      produces:
      - "application/json"
      parameters:
      - in: "path"
        name: "id"
        type: string
        required: true
      responses:
        '200':
          $ref: '#/responses/GetReferralListResponse'
        '401':
          description: "Unauthorized: Bad request or authorization details"
        '500':
          description: "Internal server error occured"
      security:
        - Bearer: []
  /v1/referral/downloadDocuments/{id}:
    get:
      tags:
       - "Referrals"
      summary: "Download all associated documents as zip file"
      description:  "Download all associated documents as zip file"
      operationId: "DownloadDocumentsAsZip"
      consumes:
         - application/json
      produces:
      - "application/json"
      parameters:
      - in: "path"
        name: "id"
        type: string
        required: true
      responses:
        '200':
          description: "response if type Content-Disposition, streaming file data to local download"
        '401':
          description: "Unauthorized: Bad request or authorization details"
        '500':
          description: "Internal server error occured"
      security:
        - Bearer: []
securityDefinitions:
  Bearer:
    type: "apiKey"
    name: "Authorization"
    in: "header"
definitions:
  GetReferralList:
    type: array
    items:
      type: object
      properties:
        referralId:
          type: string
        document:
          type: array
          items:
            type: string
        fromPlaceId:
          type: string
        toPlaceId:
          type: string
        fromClinicName:
          type: string
        toClinicName:
          type: string
        fromClinicAddress:
          type: string
        toClinicAddress:
          type: string
        comments:
          type: array
          items:
            type: object
            properties:
              time:
                type: string
              comment:
                type: string
              createdBy:
                type: string
        status:
          type: object
          properties:
            gdStatus:
              type: string
            spStatus:
              type: string
        reasons:
          type: array
          items:
            type: string
        history:
          type: array
          items:
            type: string
        tooth:
          type: array
          items:
            type: string
        createOn:
          type: string
        modifiedOn:
          type: string
        patientEmail:
          type: string
        patientFirstName:
          type: string
        patientLastName:
          type: string
        patientPhone:
          type: string
        fromEmail:
          type: string
        toEmail:
          type: string
        isDirty:
          type: boolean
          default: False
  ReferralStatus:
    type: object
    properties:
      status:
        type: object
        properties:
          gdStatus:
            type: string
          spStatus:
            type: string
  ReferralComments:
    type: object
    properties:
      comments:
        type: array
        items:
          type: object
          properties:
            time:
              type: string
            comment:
              type: string
            createdBy:
              type: string
  ReferralResponse:
    type: object
    properties:
      referralId:
        type: string
      document:
        type: array
        items:
          type: string
      fromPlaceId:
        type: string
      toPlaceId:
        type: string
      fromClinicName:
        type: string
      toClinicName:
        type: string
      fromClinicAddress:
        type: string
      toClinicAddress:
        type: string
      comments:
        type: array
        items:
          type: object
          properties:
            time:
              type: string
            comment:
              type: string
            createdBy:
              type: string
      status:
        type: object
        properties:
          gdStatus:
            type: string
          spStatus:
            type: string
      reasons:
        type: array
        items:
          type: string
      history:
        type: array
        items:
          type: string
      tooth:
        type: array
        items:
          type: string
      createOn:
        type: string
      modifiedOn:
        type: string
      patientEmail:
        type: string
      patientFirstName:
        type: string
      patientLastName:
        type: string
      patientPhone:
        type: string
      fromEmail:
        type: string
      toEmail:
        type: string
      isDirty:
        type: boolean
        default: False
...
responses:
  GetReferralListResponse:
    description: "List of referrals"
    schema:
      $ref: '#/definitions/GetReferralList'
...
  CreateReferralResponse:
    description: On successful referral creation payload containing unique referral id is returned and emails are send to patient and clinic for whom referral is created
    schema:
      $ref: '#/definitions/ReferralResponse'